checked of pages
My map page is still just the map image at this point. I found a way to create an actual gps map (by digging through the source code of the UM's website, but it might be biting off more than we can chew. In the mean time, I will try to get a zoom going on the static image.